Reunite checkpoint and backup core logic
